refactor: useModalAccessibility 훅 추출 및 inert 적용 방식 개선 #117
